reduce the shapes after type-checking
The previous approach to shape reduction, which use the Shape.app
smart constructor during module type-checking, would perform shape
reduction under functor abstractions -- strong reductions.

Strong reductions are the case that blows up on some interesting
examples, including Irmin: under functor abstractions, shapes of
module type may not reduce to structures (but neutral terms containing
free variables), and the projections
  {f = S.f; g = S.g; ...}
that correspond to the translation of module signature coercions
  (M : sig type f type g end)
may generate arbitrary many copies of the non-trivial shape S>

The approach in this commit is to reduce shapes for the whole
module (treating references to other compilation units as free
variables; this remains a "local" reduction), but very weakly: without
going under functor abstractions, and in fact without reducing functor
applications either.

In particular, the production of irmin.cmt, a known-bad case, does not
blow up.

We added a new example in the testsuite that produces an
exponentially-large shape, but can still be computed in a short amount
of time.

Note: existing shape tests do *not* pass with this commit, as we
reduce less aggrssively than before. Follow-up commits are going to
implement a stronger reduction stragegy.