memoize reductions
This change has a radical effect on performance, it completely
eliminates the redundancies inherent in som examples, both in
time (avoiding duplicated computations) and
space (reusing physically-identical results instead of copies of the
same results).

At this point, irmin.cmt takes 0.5s to produce a 2.5Mio result; this
is fast and small.